+++ mix.exs
@@ -46,8 +46,7 @@ defmodule CCC.Mixfile do
   #
   # Type `mix help compile.app` for more information
   def application do
-    [applications: [:logger],
-     mod: {CCC, []}]
+    [applications: [:logger]]
   end
 
   # Dependencies can be Hex packages:
